// src/api/generated/models/user-profile-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gia tot example
 * The gia tot API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UserProfileDto
 */
export interface UserProfileDto {
    /**
     * Email người dùng
     * @type {string}
     * @memberof UserProfileDto
     */
    'email': string;
    /**
     * Danh sách role của người dùng
     * @type {Array<string>}
     * @memberof UserProfileDto
     */
    'roles': Array<string>;
    /**
     * Danh sách permission của người dùng
     * @type {Array<string>}
     * @memberof UserProfileDto
     */
    'permissions': Array<string>;
}